network: bridge: create fakedev in mounted directory
Bridge device plugin creates fake device that is mounted to containers
requesting bridge resource. In case device plugin is running in a pod,
this fake device is created inside its file-system and therefore it is
not visible for host system. With this change, we create fake device
in mounted /var/run directory.

Signed-off-by: Petr Horacek <phoracek@redhat.com>